Do not dispatch signals to workers without a valid pid.
thanks rille.

+++ src/worker.c
@@ -271,6 +271,10 @@ kore_worker_dispatch_signal(int sig)
 
 	for (idx = 0; idx < worker_count; idx++) {
 		kw = WORKER(idx);
+
+		if (kw->pid == -1 || kw->pid == 0)
+			continue;
+
 		if (kill(kw->pid, sig) == -1) {
 			kore_debug("kill(%d, %d): %s", kw->pid, sig, errno_s);
 		}
